+++ src/IUserNameResolver.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+
+namespace prmonitor;
+
+internal interface IUserNameResolver
+{
+    Task<string> ResolveUsernameForLogin(string login);
+}
